splitted utils file into 2 files, also make small cleanup
