--- frontend/src/loaders/pageNotFoundLoader.ts
@@ -1,10 +1,0 @@
-import authorization from "../store/authorization";
-import { redirect } from "react-router-dom";
-
-export default function pageNotFoundLoader() {
-  if (!authorization.get()) {
-    console.warn("User is not authorized: redirecting...");
-    return redirect("/authorization");
-  }
-  return "";
-}
